Services/Dossier/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
slack_sdk>=3.30.0
anthropic>=0.35.0
google-cloud-storage>=2.18.2
google-cloud-secret-manager>=2.21.0
httpx>=0.27.0
python-dotenv>=1.0.0
